tracing: 新しいDatadogトレース機能を追加し、トレース設定を改善
